app/pyproject.toml
[project.urls]
"Homepage" = "https://github.com/ThePyrotechnic/simplechat"
"Bug Tracker" = "https://github.com/ThePyrotechnic/simplechat/issues"

[tool.poetry]
name = "simplechat"
version = "0.1.0"
description = "Chat with AI"
authors = ["Michael Manis <michaelmanis@tutanota.com>"]
license = "AGPL-3.0-only"
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.34.150"
fastapi = {extras = ["all"], version = "^0.111.1"}
PyJWT = {extras = ["crypto"], version = "^2.8.0"}
python-dotenv = "^1.0.1"
uvicorn = "^0.30.3"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
httpx = "^0.27.0"
pytest = "^8.3.2"


[tool.poetry.scripts]
simplechat = "src.simplechat.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
